fix: satisfy backend-size limits for session context PR
Extract configuredWorkstationExecutor helper and move session-context
executor tests to a focused file so CI backend-size checks pass.
